// src/Strata.DSS.CostAccounting.Web/dss.costaccounting/src/department-categorization/enums/ExceptionNameEnum.ts
export enum ExceptionNameEnum {
  RevenueToOverhead = 'Revenue to Overhead',
  RevenueToExcluded = 'Revenue to Excluded',
  OverheadToRevenue = 'Overhead to Revenue',
  OverheadToExcluded = 'Overhead to Excluded',
  ExcludedToOverhead = 'Excluded to Overhead',
  ExcludedToRevenue = 'Excluded to Revenue',
  //claims
  IncludedToExcluded = 'Included to Excluded',
  ExcludedToIncluded = 'Excluded to Included'
}